fix: add browser testing
